move AggregationDistinctToken
